Supports decimal review ratings and enhances score display
Switches review ratings from integers to doubles to allow decimal scores, improving rating accuracy and expressiveness. Updates UI to show average and individual reviewer scores with new visual styles and formatting for clarity. Refines CSS for responsive and accessible score presentation.

+++ BoardGameMondays/Core/MemberReview.cs
@@ -2,7 +2,7 @@ namespace BoardGameMondays.Core;
 
 public sealed class MemberReview : Review
 {
-    public MemberReview(BgmMember reviewer, int rating, string description, IEnumerable<Comment>? comments = null)
+    public MemberReview(BgmMember reviewer, double rating, string description, IEnumerable<Comment>? comments = null)
     {
         Reviewer = reviewer;
         Rating = rating;
@@ -12,7 +12,7 @@ public sealed class MemberReview : Review
 
     public override BgmMember Reviewer { get; }
 
-    public override int Rating { get; }
+    public override double Rating { get; }
 
     public override string Description { get; }
